move global baserom folder detection to a new file that the makefile can use
